refactor: styledCard 중복 제거를 위한 children 컴포넌트 구현
