+++ client/src/features/Event/New/components/TemplateList.tsx
@@ -1,11 +1,12 @@
-import styled from '@emotion/styled';
+import { css } from '@emotion/react';
 
 import { useDeleteTemplate } from '@/api/mutations/useDeleteTemplate';
-import { Card } from '@/shared/components/Card';
 import { Flex } from '@/shared/components/Flex';
 import { IconButton } from '@/shared/components/IconButton';
 import { Text } from '@/shared/components/Text';
 import { theme } from '@/shared/styles/theme';
+
+import { SelectableCard } from './SelectableCard';
 
 type Template = {
   templateId: number;
@@ -16,10 +17,6 @@ type TemplateListProps = {
   templates: Template[];
   selectedId: number;
   onSelectTemplate: (templateId: number) => void;
-};
-
-type StyledCardProps = {
-  isSelected: boolean;
 };
 
 export const TemplateList = ({ templates, selectedId, onSelectTemplate }: TemplateListProps) => {
@@ -38,29 +35,52 @@ export const TemplateList = ({ templates, selectedId, onSelectTemplate }: Templa
       <Text type="Body" weight="regular" color={theme.colors.gray600}>
         저장된 템플릿이 여기에 표시됩니다.
       </Text>
-      {templates?.map((template) => {
-        const isSelected = selectedId === template.templateId;
 
-        return (
-          <StyledCard
-            key={template.templateId}
-            isSelected={isSelected}
-            onClick={() => onSelectTemplate(template.templateId)}
-          >
-            <Flex justifyContent="space-between" alignItems="center">
-              <Text type="Body" weight="medium" color={theme.colors.gray900}>
-                {template.title}
-              </Text>
-              <IconButton
-                name="delete"
-                size={20}
-                color="gray"
-                onClick={() => handleDeleteTemplate(template.templateId)}
-              />
-            </Flex>
-          </StyledCard>
-        );
-      })}
+      <Flex
+        dir="column"
+        gap="12px"
+        css={css`
+          max-height: 200px;
+          overflow-y: auto;
+
+          &::-webkit-scrollbar {
+            width: 6px;
+          }
+
+          &::-webkit-scrollbar-thumb {
+            background: ${theme.colors.gray300};
+            border-radius: 3px;
+
+            &:hover {
+              background: ${theme.colors.gray400};
+            }
+          }
+        `}
+      >
+        {templates?.map((template) => {
+          const isSelected = selectedId === template.templateId;
+
+          return (
+            <SelectableCard
+              key={template.templateId}
+              isSelected={isSelected}
+              onClick={() => onSelectTemplate(template.templateId)}
+            >
+              <Flex justifyContent="space-between" alignItems="center">
+                <Text type="Body" weight="medium" color={theme.colors.gray900}>
+                  {template.title}
+                </Text>
+                <IconButton
+                  name="delete"
+                  size={20}
+                  color="gray"
+                  onClick={() => handleDeleteTemplate(template.templateId)}
+                />
+              </Flex>
+            </SelectableCard>
+          );
+        })}
+      </Flex>
 
       {templates?.length === 0 && (
         <Flex alignItems="center" justifyContent="center" padding="40px 0">
@@ -72,15 +92,3 @@ export const TemplateList = ({ templates, selectedId, onSelectTemplate }: Templa
     </Flex>
   );
 };
-
-const StyledCard = styled(Card)<StyledCardProps>`
-  cursor: pointer;
-  padding: 16px;
-  background-color: ${({ isSelected }) =>
-    isSelected ? theme.colors.primary50 : theme.colors.white};
-  border-radius: 8px;
-  transition: all 0.2s ease;
-  &:hover {
-    background-color: ${theme.colors.primary50};
-  }
-`;
